services/personality_layer/requirements.txt
torch==2.7.1
faiss-cpu==1.11.0
requests==2.32.4
openai==1.86.0
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.0.0